requirements.txt
fastapi
uvicorn[standard]
python-dotenv
mysql-connector-python
pydantic
supabase==2.15.3 